Allow convert to record with equals, hashCode, and toString methods
- change prereq for convert to record to be Java 16 to avoid
  formatting issues
- add new logic to ConvertToRecordRefactoring to copy equals,
  hashCode, and toString methods to resultant record
- move tests to AssistQuickFixTests16 and add new test for
  moving methods
- fix various formatting issues that were built-in to the tests when
  run at Java 14
- fixes #2795
