+++ org.eclipse.jdt.core.manipulation/core extension/org/eclipse/jdt/internal/corext/refactoring/code/ConvertToRecordRefactoring.java
@@ -45,6 +45,7 @@ import org.eclipse.jdt.core.dom.ASTVisitor;
 import org.eclipse.jdt.core.dom.AbstractTypeDeclaration;
 import org.eclipse.jdt.core.dom.AnonymousClassDeclaration;
 import org.eclipse.jdt.core.dom.Assignment;
+import org.eclipse.jdt.core.dom.BodyDeclaration;
 import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.Expression;
 import org.eclipse.jdt.core.dom.ExpressionStatement;
@@ -60,15 +61,18 @@ import org.eclipse.jdt.core.dom.MethodDeclaration;
 import org.eclipse.jdt.core.dom.MethodInvocation;
 import org.eclipse.jdt.core.dom.Modifier;
 import org.eclipse.jdt.core.dom.NodeFinder;
+import org.eclipse.jdt.core.dom.PrimitiveType;
 import org.eclipse.jdt.core.dom.RecordDeclaration;
 import org.eclipse.jdt.core.dom.ReturnStatement;
 import org.eclipse.jdt.core.dom.SimpleName;
+import org.eclipse.jdt.core.dom.SimpleType;
 import org.eclipse.jdt.core.dom.SingleVariableDeclaration;
 import org.eclipse.jdt.core.dom.Statement;
 import org.eclipse.jdt.core.dom.TypeDeclaration;
 import org.eclipse.jdt.core.dom.VariableDeclarationFragment;
 import org.eclipse.jdt.core.dom.rewrite.ASTRewrite;
 import org.eclipse.jdt.core.dom.rewrite.TargetSourceRangeComputer;
+import org.eclipse.jdt.core.refactoring.CompilationUnitChange;
 import org.eclipse.jdt.core.refactoring.descriptors.ConvertToRecordDescriptor;
 import org.eclipse.jdt.core.refactoring.descriptors.JavaRefactoringDescriptor;
 import org.eclipse.jdt.core.search.IJavaSearchConstants;
@@ -81,7 +85,6 @@ import org.eclipse.jdt.internal.core.manipulation.JavaElementLabelsCore;
 import org.eclipse.jdt.internal.core.manipulation.util.BasicElementLabels;
 import org.eclipse.jdt.internal.core.refactoring.descriptors.RefactoringSignatureDescriptorFactory;
 import org.eclipse.jdt.internal.corext.codemanipulation.GetterSetterUtil;
-import org.eclipse.jdt.internal.corext.dom.ASTFlattener;
 import org.eclipse.jdt.internal.corext.dom.ASTNodes;
 import org.eclipse.jdt.internal.corext.dom.Bindings;
 import org.eclipse.jdt.internal.corext.dom.Selection;
@@ -113,6 +116,7 @@ public class ConvertToRecordRefactoring extends Refactoring {
 	Map<IVariableBinding, IMethodBinding> fGetterMap= new HashMap<>();
 	TextChangeManager fChangeManager;
 	CompilationUnitRewrite fBaseCURewrite;
+	List<MethodDeclaration> fMethodsToCopy= new ArrayList<>();
 
 
 	public ConvertToRecordRefactoring(ICompilationUnit unit, CompilationUnit node, int selectionStart, int selectionLength) {
@@ -238,9 +242,6 @@ public class ConvertToRecordRefactoring extends Refactoring {
 					}
 				}
 
-				if (methods.length > fields.length + 1) {
-					return RefactoringStatus.createFatalErrorStatus(RefactoringCoreMessages.ConvertToRecordRefactoring_not_simple_case);
-				}
 				class VisitException extends RuntimeException {
 					private static final long serialVersionUID= 1L;
 
@@ -286,7 +287,36 @@ public class ConvertToRecordRefactoring extends Refactoring {
 									throw new VisitException(RefactoringCoreMessages.ConvertToRecordRefactoring_nonstandard_constructor);
 								}
 							} else {
-								if (statements.size() != 1 || !(statements.get(0) instanceof ReturnStatement retStmt)) {
+								if (node.getName().getFullyQualifiedName().equals("equals")) { //$NON-NLS-1$
+									if (Modifier.isPublic(node.getModifiers())
+											&& node.getReturnType2() instanceof PrimitiveType primitiveType
+											&& primitiveType.getPrimitiveTypeCode() == PrimitiveType.BOOLEAN
+											&& node.parameters().size() == 1
+											&& ((SingleVariableDeclaration)node.parameters().get(0)).getType().resolveBinding() instanceof ITypeBinding typeBinding
+											&& typeBinding.getQualifiedName().equals("java.lang.Object")) { //$NON-NLS-1$
+										fMethodsToCopy.add(node);
+									} else {
+										throw new VisitException(RefactoringCoreMessages.ConvertToRecordRefactoring_not_simple_case);
+									}
+								} else if (node.getName().getFullyQualifiedName().equals("hashCode")) { //$NON-NLS-1$
+									if (Modifier.isPublic(node.getModifiers())
+											&& node.getReturnType2() instanceof PrimitiveType primitiveType
+											&& primitiveType.getPrimitiveTypeCode() == PrimitiveType.INT
+											&& node.parameters().size() == 0) {
+										fMethodsToCopy.add(node);
+									} else {
+										throw new VisitException(RefactoringCoreMessages.ConvertToRecordRefactoring_not_simple_case);
+									}
+								} else if (node.getName().getFullyQualifiedName().equals("toString")) { //$NON-NLS-1$
+									if (Modifier.isPublic(node.getModifiers())
+											&& node.getReturnType2() instanceof SimpleType simpleType
+											&& simpleType.getName().getFullyQualifiedName().equals("String") //$NON-NLS-1$
+											&& node.parameters().size() == 0) {
+										fMethodsToCopy.add(node);
+									} else {
+										throw new VisitException(RefactoringCoreMessages.ConvertToRecordRefactoring_not_simple_case);
+									}
+							} else if (statements.size() != 1 || !(statements.get(0) instanceof ReturnStatement retStmt)) {
 									throw new VisitException(RefactoringCoreMessages.ConvertToRecordRefactoring_not_implicit_getter);
 								} else {
 									Expression exp= retStmt.getExpression();
@@ -324,6 +354,9 @@ public class ConvertToRecordRefactoring extends Refactoring {
 							return null;
 						}
 					});
+					if (methods.length > fields.length + fMethodsToCopy.size() + 1) {
+						return RefactoringStatus.createFatalErrorStatus(RefactoringCoreMessages.ConvertToRecordRefactoring_not_simple_case);
+					}
 				} catch (VisitException e) {
 					return RefactoringStatus.createFatalErrorStatus(e.getMessage());
 				}
@@ -414,7 +447,10 @@ public class ConvertToRecordRefactoring extends Refactoring {
 			ArrayList<Change> changes= new ArrayList<>();
 			createNewRecord();
 			updateReferences(pm);
-			fChangeManager.manage(typeCU, fBaseCURewrite.createChange(true, pm));
+
+			CompilationUnitChange rewriteChange= fBaseCURewrite.createChange(true, pm);
+			System.out.println(rewriteChange.getEdit());
+			fChangeManager.manage(typeCU, rewriteChange);
 			changes.addAll(Arrays.asList(fChangeManager.getAllChanges()));
 			final Map<String, String> arguments= new HashMap<>();
 			String project= null;
@@ -452,44 +488,30 @@ public class ConvertToRecordRefactoring extends Refactoring {
 		});
 
 		RecordDeclaration newRecordDeclaration= null;
-		if (fTypeDeclaration.getJavadoc() instanceof Javadoc javadoc) {
-			StringBuilder builder= new StringBuilder();
-			builder.append(ASTFlattener.asString(javadoc));
-			builder.append("\n"); //$NON-NLS-1$
-			List<IExtendedModifier> modifiers= fTypeDeclaration.modifiers();
-			for (IExtendedModifier modifier : modifiers) {
-				if (!(modifier instanceof Modifier mod) || !mod.isFinal()) {
-					builder.append(modifier.toString());
-					builder.append(" "); //$NON-NLS-1$
-				}
+		newRecordDeclaration= ast.newRecordDeclaration();
+		newRecordDeclaration.setName((SimpleName) rewrite.createCopyTarget(fTypeDeclaration.getName()));
+		if (fTypeDeclaration.getJavadoc() != null) {
+			Javadoc newJavaDoc= (Javadoc) rewrite.createCopyTarget(fTypeDeclaration.getJavadoc());
+			newRecordDeclaration.setJavadoc(newJavaDoc);
+		}
+		List<IExtendedModifier> modifiers= fTypeDeclaration.modifiers();
+		List<IExtendedModifier> recordModifiers= newRecordDeclaration.modifiers();
+		for (IExtendedModifier modifier : modifiers) {
+			if (!(modifier instanceof Modifier mod) || !mod.isFinal()) {
+				IExtendedModifier newModifier= (IExtendedModifier) rewrite.createCopyTarget((ASTNode)modifier);
+				recordModifiers.add(newModifier);
 			}
-			builder.append("record " + fTypeDeclaration.getName().getFullyQualifiedName() + " ("); //$NON-NLS-1$ //$NON-NLS-2$
-			List<SingleVariableDeclaration> parameters= fConstructor.parameters();
-			String separator= ""; //$NON-NLS-1$
-			for (SingleVariableDeclaration parameter : parameters) {
-				builder.append(separator);
-				builder.append(ASTFlattener.asString(parameter));
-				separator= ", "; //$NON-NLS-1$
-			}
-			builder.append(") {}"); //$NON-NLS-1$
-			newRecordDeclaration= (RecordDeclaration) rewrite.createStringPlaceholder(builder.toString(), ASTNode.RECORD_DECLARATION);
-		} else {
-			newRecordDeclaration= ast.newRecordDeclaration();
-			newRecordDeclaration.setName((SimpleName) rewrite.createCopyTarget(fTypeDeclaration.getName()));
-			List<IExtendedModifier> modifiers= fTypeDeclaration.modifiers();
-			List<IExtendedModifier> recordModifiers= newRecordDeclaration.modifiers();
-			for (IExtendedModifier modifier : modifiers) {
-				if (!(modifier instanceof Modifier mod) || !mod.isFinal()) {
-					IExtendedModifier newModifier= (IExtendedModifier) rewrite.createCopyTarget((ASTNode)modifier);
-					recordModifiers.add(newModifier);
-				}
-			}
-			List<SingleVariableDeclaration> components= newRecordDeclaration.recordComponents();
-			List<SingleVariableDeclaration> parameters= fConstructor.parameters();
-			for (SingleVariableDeclaration parameter : parameters) {
-				SingleVariableDeclaration newSingleVariableDeclaration= (SingleVariableDeclaration) rewrite.createCopyTarget(parameter);
-				components.add(newSingleVariableDeclaration);
-			}
+		}
+		List<SingleVariableDeclaration> components= newRecordDeclaration.recordComponents();
+		List<SingleVariableDeclaration> parameters= fConstructor.parameters();
+		for (SingleVariableDeclaration parameter : parameters) {
+			SingleVariableDeclaration newSingleVariableDeclaration= (SingleVariableDeclaration) rewrite.createCopyTarget(parameter);
+			components.add(newSingleVariableDeclaration);
+		}
+		List<BodyDeclaration> bodyDeclarations= newRecordDeclaration.bodyDeclarations();
+		for (MethodDeclaration methodDecl : fMethodsToCopy) {
+			MethodDeclaration newMethodDeclaration= (MethodDeclaration) rewrite.createCopyTarget(methodDecl);
+			bodyDeclarations.add(newMethodDeclaration);
 		}
 		ASTNodes.replaceButKeepComment(rewrite, fTypeDeclaration, newRecordDeclaration, null);
 	}
