+++ org.eclipse.jdt.ui.tests.refactoring/resources/ConvertToRecord/canConvert/A_test0_out.java
@@ -1,3 +1,4 @@
 package p;
 // Class A
-public record A (int a,String b) {}
+public record A(int a, String b) {
+}
