+++ org.eclipse.jdt.ui.tests.refactoring/resources/ConvertToRecord/canConvert/A_test4_in.java
@@ -1,0 +1,25 @@
+package p;
+// Class A
+public final class A {
+
+	private final int a;
+	private final String b;
+
+	public A(int a, String b) {
+		this.a= a;
+		this.b= b;
+	}
+
+	public int getA() {
+		return a;
+	}
+
+	public String getB() {
+		return b;
+	}
+
+	@Override
+	public String toString() {
+		return "toString";
+	}
+}
