+++ org.eclipse.jdt.ui.tests.refactoring/resources/ConvertToRecord/canConvert/A_test4_out.java
@@ -1,0 +1,8 @@
+package p;
+// Class A
+public record A(int a, String b) {
+	@Override
+	public String toString() {
+		return "toString";
+	}
+}
