+++ org.eclipse.jdt.ui.tests/ui/org/eclipse/jdt/ui/tests/quickfix/AssistQuickFixTest14.java
@@ -29,19 +29,17 @@ import org.eclipse.jdt.core.IPackageFragment;
 import org.eclipse.jdt.core.IPackageFragmentRoot;
 
 import org.eclipse.jdt.internal.corext.fix.FixMessages;
-import org.eclipse.jdt.internal.corext.refactoring.RefactoringCoreMessages;
 
 import org.eclipse.jdt.ui.tests.core.rules.Java14ProjectTestSetup;
 import org.eclipse.jdt.ui.tests.core.rules.ProjectTestSetup;
 import org.eclipse.jdt.ui.text.java.IInvocationContext;
 import org.eclipse.jdt.ui.text.java.IJavaCompletionProposal;
 import org.eclipse.jdt.ui.text.java.correction.CUCorrectionProposal;
-import org.eclipse.jdt.ui.text.java.correction.ChangeCorrectionProposal;
 
 public class AssistQuickFixTest14 extends QuickFixTest {
 
 	@Rule
-    public ProjectTestSetup projectSetup = new Java14ProjectTestSetup(true);
+    public ProjectTestSetup projectSetup = new Java14ProjectTestSetup(false);
 
 	private IJavaProject fJProject1;
 
@@ -802,737 +800,5 @@ public class AssistQuickFixTest14 extends QuickFixTest {
 		assertProposalDoesNotExist(proposals, FixMessages.SwitchExpressionsFix_convert_to_switch_expression);
 	}
 
-	@Test
-	public void testConvertToRecord1() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final int a;
-					private final String b;
-
-					public Cls(int a, String b) {
-						this.a= a;
-						this.b= b;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-				}
-				""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("a;");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 1);
-		assertNoErrors(ctx);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		ChangeCorrectionProposal proposal= (ChangeCorrectionProposal) proposals.get(0);
-		proposal.apply();
-		String preview= cu.getBuffer().getContents();
-
-		String expected= """
-				package test;
-
-				public record Cls (int a,String b) {}
-				""";
-
-		assertEqualString(expected, preview);
-	}
-
-	@Test
-	public void testConvertToRecord2() throws Exception {
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final int a;
-					private final String b;
-
-					public Cls(int a, String b) {
-						this.a= a;
-						this.b= b;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-				}
-				""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		String str2= """
-				package test;
-
-				public class Cls2 {
-					public void foo() {
-						Cls cls= new Cls(3, "abc");
-						System.out.println(cls.getA());
-						System.out.println(cls.getB());
-					}
-				}
-				""";
-		ICompilationUnit cu2= pack.createCompilationUnit("Cls2.java", str2, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		assertNoErrors(ctx);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		ChangeCorrectionProposal proposal= (ChangeCorrectionProposal) proposals.get(0);
-		proposal.apply();
-		String preview= cu.getBuffer().getContents();
-
-		String expected= """
-				package test;
-
-				public record Cls (int a,String b) {}
-				""";
-
-		assertEqualString(expected, preview);
-
-		String preview2= cu2.getBuffer().getContents();
-
-		String expected2= """
-				package test;
-
-				public class Cls2 {
-					public void foo() {
-						Cls cls= new Cls(3, "abc");
-						System.out.println(cls.a());
-						System.out.println(cls.b());
-					}
-				}
-				""";
-
-		assertEqualString(expected2, preview2);
-	}
-
-	@Test
-	public void testConvertToRecord3() throws Exception {
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					// Inner class
-					public static class Inner {
-						private final int a;
-						private final String b;
-
-						public Inner(int a, String b) {
-							this.a= a;
-							this.b= b;
-						}
-
-						public int getA() {
-							return a;
-						}
-
-						public String getB() {
-							return b;
-						}
-					}
-				}
-				""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		String str2= """
-				package test;
-
-				public class Cls2 {
-					public void foo() {
-						Cls.Inner cls= new Cls.Inner(3, "abc");
-						System.out.println(cls.getA());
-						System.out.println(cls.getB());
-					}
-				}
-				""";
-		ICompilationUnit cu2= pack.createCompilationUnit("Cls2.java", str2, false, null);
-
-		int index= str1.indexOf("Inner {");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 5);
-		assertNoErrors(ctx);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		ChangeCorrectionProposal proposal= (ChangeCorrectionProposal) proposals.get(0);
-		proposal.apply();
-		String preview= cu.getBuffer().getContents();
-
-		String expected= """
-				package test;
-
-				public class Cls {
-					// Inner class
-					public static record Inner (int a,String b) {}
-				}
-				""";
-
-		assertEqualString(expected, preview);
-
-		String preview2= cu2.getBuffer().getContents();
-
-		String expected2= """
-				package test;
-
-				public class Cls2 {
-					public void foo() {
-						Cls.Inner cls= new Cls.Inner(3, "abc");
-						System.out.println(cls.a());
-						System.out.println(cls.b());
-					}
-				}
-				""";
-
-		assertEqualString(expected2, preview2);
-	}
-
-	@Test
-	public void testConvertToRecord4() throws Exception {
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					/**
-					 * Class Inner
-					 */
-					private final class Inner {
-						private final int a;
-						private final String b;
-						private double c;
-
-						public Inner(int a, String b, double c) {
-							this.a= a;
-							this.b= b;
-							this.c= c;
-						}
-
-						public int getA() {
-							return a;
-						}
-
-						public String getB() {
-							return b;
-						}
-
-						public double getC() {
-							return c;
-						}
-					}
-					public void foo() {
-						Inner inner= new Inner(1, "comment", 4.3);
-						System.out.println(inner.getA());
-						System.out.println(inner.getB());
-						System.out.println(inner.getC());
-					}
-				}
-				""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Inner(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 5);
-		assertNoErrors(ctx);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		ChangeCorrectionProposal proposal= (ChangeCorrectionProposal) proposals.get(0);
-		proposal.apply();
-		String preview= cu.getBuffer().getContents();
-
-		String expected= """
-				package test;
-
-				public class Cls {
-					/**\s
-					 * Class Inner
-					 */
-					private record Inner (int a, String b, double c) {}
-					public void foo() {
-						Inner inner= new Inner(1, "comment", 4.3);
-						System.out.println(inner.a());
-						System.out.println(inner.b());
-						System.out.println(inner.c());
-					}
-				}
-				""";
-
-		assertEqualString(preview, expected);
-	}
-
-	@Test
-	public void testConvertToRecord5() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				/* Class Cls */
-				public class Cls extends Object {
-					private final int a;
-					private final String b;
-
-					public Cls(int a, String b) {
-						this.a= a;
-						this.b= b;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-				}
-				""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		assertNoErrors(ctx);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		ChangeCorrectionProposal proposal= (ChangeCorrectionProposal) proposals.get(0);
-		proposal.apply();
-		String preview= cu.getBuffer().getContents();
-
-		String expected= """
-				package test;
-
-				/* Class Cls */
-				public record Cls (int a,String b) {}
-				""";
-
-		assertEqualString(expected, preview);
-	}
-
-	@Test
-	public void testNoConvertToRecord1() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final int a;
-					private final String b;
-					private double c;
-
-					public Cls(int a, String b, double c) {
-						this.a= a;
-						this.b= b;
-						this.c= c;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-				}
-			""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
-
-	@Test
-	public void testNoConvertToRecord2() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final int a;
-					private final String b;
-					private double c = 2.4;;
-
-					public Cls(int a, String b) {
-						this.a= a;
-						this.b= b;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-				}
-			""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
-
-	@Test
-	public void testNoConvertToRecord3() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final static int a;
-					private final String b;
-					private double c;;
-
-					public Cls(int a, String b, double c) {
-						this.a= a;
-						this.b= b;
-						this.c= c;
-					}
-
-					public static int getAValue() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-
-					public double getC() {
-						return c;
-					}
-				}
-			""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
-
-	@Test
-	public void testNoConvertToRecord4() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final int a;
-					private final String b;
-					private double c;;
-
-					public Cls(int a, String b, double c) {
-						this.a= a;
-						this.b= b;
-						this.c= c;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-
-					public double getC() {
-						return c;
-					}
-
-					public int getSum() {
-						return a + b.length();
-					}
-				}
-			""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
-
-	@Test
-	public void testNoConvertToRecord5() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					private final int a;
-					private final String b;
-					public double c;;
-
-					public Cls(int a, String b, double c) {
-						this.a= a;
-						this.b= b;
-						this.c= c;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-
-					public double getC() {
-						return c;
-					}
-				}
-			""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
-
-	@Test
-	public void testNoConvertToRecord6() throws Exception { // https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2681
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				class K {
-				}
-				public class Cls extends K {
-					private final int a;
-					private final String b;
-					private double c;;
-
-					public Cls(int a, String b, double c) {
-						this.a= a;
-						this.b= b;
-						this.c= c;
-					}
-
-					public int getA() {
-						return a;
-					}
-
-					public String getB() {
-						return b;
-					}
-
-					public double getC() {
-						return c;
-					}
-				}
-			""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Cls(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 3);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
-
-	@Test
-	public void testNoConvertToRecord7() throws Exception {
-		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
-		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
-		JavaProjectHelper.set14CompilerOptions(fJProject1, false);
-		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
-
-		String str= """
-			module test {
-			}
-			""";
-		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
-		def.createCompilationUnit("module-info.java", str, false, null);
-
-		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
-		String str1= """
-				package test;
-
-				public class Cls {
-					/**
-					 * Class Inner
-					 */
-					private class Inner {
-						private final int a;
-						private final String b;
-						private double c;
-
-						public Inner(int a, String b, double c) {
-							this.a= a;
-							this.b= b;
-							this.c= c;
-						}
-
-						public int getA() {
-							return a;
-						}
-
-						public String getB() {
-							return b;
-						}
-
-						public double getC() {
-							return c;
-						}
-					}
-					private class Inner2 extends Inner {
-						public Inner2() {
-							super(2, "blah", 5.2);
-						}
-					}
-					public void foo() {
-						Inner inner= new Inner(1, "comment", 4.3);
-						System.out.println(inner.getA());
-						System.out.println(inner.getB());
-						System.out.println(inner.getC());
-					}
-				}
-				""";
-		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
-
-		int index= str1.indexOf("Inner(");
-		IInvocationContext ctx= getCorrectionContext(cu, index, 5);
-		assertNoErrors(ctx);
-		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
-		assertProposalDoesNotExist(proposals, RefactoringCoreMessages.ConvertToRecordRefactoring_name);
-	}
 }
